+++ backend/canisters/notifications/api/src/updates/c2c_push_notification.rs
@@ -2,13 +2,12 @@ use candid::CandidType;
 use serde::{Deserialize, Serialize};
 use serde_bytes::ByteBuf;
 use std::fmt::{Debug, Formatter};
-use types::{CanisterId, UserId};
+use types::UserId;
 
 #[derive(CandidType, Serialize, Deserialize)]
 pub struct Args {
     pub sender: Option<UserId>,
     pub recipients: Vec<UserId>,
-    pub authorizer: Option<CanisterId>,
     pub notification_bytes: ByteBuf,
 }
 
@@ -19,7 +18,6 @@ impl Debug for Args {
         f.debug_struct("Args")
             .field("sender", &self.sender)
             .field("recipients", &self.recipients)
-            .field("authorizer", &self.authorizer)
             .field("notification_bytes_length", &self.notification_bytes.len())
             .finish()
     }