+++ backend/canisters/notifications/impl/src/model/authorized_principals.rs
@@ -1,34 +1,20 @@
 use candid::Principal;
 use serde::{Deserialize, Serialize};
 use std::collections::HashSet;
-use types::CanisterId;
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, Default)]
 pub struct AuthorizedPrincipals {
-    authorizers: HashSet<CanisterId>,
     authorized: HashSet<Principal>,
     blocked: HashSet<Principal>,
 }
 
 impl AuthorizedPrincipals {
-    pub fn new(authorizers: HashSet<CanisterId>) -> AuthorizedPrincipals {
-        AuthorizedPrincipals {
-            authorizers,
-            authorized: HashSet::new(),
-            blocked: HashSet::new(),
-        }
-    }
-
     pub fn add_principal(&mut self, principal: Principal, authorized: bool) {
         if authorized {
             self.authorized.insert(principal);
         } else {
             self.blocked.insert(principal);
         }
-    }
-
-    pub fn is_authorizer(&self, canister_id: &CanisterId) -> bool {
-        self.authorizers.contains(canister_id)
     }
 
     pub fn can_push_notifications(&self, principal: &Principal) -> Option<bool> {