+++ backend/canisters/registry/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Changed
+
+- Remove LocalGroupIndex from Notifications canister init args ([#7861](https://github.com/open-chat-labs/open-chat/pull/7861))
+
 ## [[2.0.1723](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1723-registry)] - 2025-05-02
 
 ### Changed