+++ backend/canisters/user/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 
+- Avoid having to specify `authorizer` when pushing notifications ([#7861](https://github.com/open-chat-labs/open-chat/pull/7861))
 - Migrate `bot_context` to `sender_context` ([#7870](https://github.com/open-chat-labs/open-chat/pull/7870))
 
 ### Fixed